# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YAMAGUCHI Demo API",
        "description": "Документация к API для тестового задания YAMAGUCHI",
        "contact": {
            "email": "nickgaidai@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "YAMAGUCHI API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрирует нового пользователя",
                "description": "Регистрирует нового пользователя",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Адрес электронной почты",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Подтверждение пароля",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь зарегистрирован"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Аутентифицирует пользователя и выдает JWT-токен",
                "description": "Аутентифицирует пользователя и выдает JWT-токен",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Адрес электронной почты",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный вход"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Возвращает авторизованного пользователя",
                "description": "Возвращает авторизованного пользователя",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Деаутентифицирует пользователя (Сбрасывает токен)",
                "description": "Деаутентифицирует пользователя (Сбрасывает токен)",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Запрос выполнен"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Идентификатор пользователя",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Адрес электронной почты",
                        "type": "string",
                        "example": "test@example.com"
                    },
                    "email_verified_at": {
                        "title": "Email Verified at",
                        "description": "Дата регистрации",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45",
                        "nullable": true
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Дата регистрации",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Эндпоинты для аутентификации"
        }
    ]
}